Implementación temporal del plugin security y las ACL's, deben modificar un usuario y ponerle como userrole "ROLE_SUDO" para la administración total de la aplicación

+++ app/controllers/FieldController.php
@@ -11,6 +11,10 @@ class FieldController extends ControllerBase
 	 */
 	public function indexAction()
 	{
+		$r = $this->verifyAcl('field', 'index', '');
+		if ($r)
+			return $r;
+		
 		$idbase = 0;
 		$dbases = $this->user->account->dbases;
 		
@@ -36,6 +40,10 @@ class FieldController extends ControllerBase
 	 */
 	public function insertAction()
 	{
+		$r = $this->verifyAcl('field', 'insert', '');
+		if ($r)
+			return $r;
+		
 		$idBase = $this->request->get('idDbase', 'int', 0);
 		
 		$this->view->disable();
@@ -59,6 +67,10 @@ class FieldController extends ControllerBase
 	 */
 	public function updateAction($idCustomField)
 	{
+		$r = $this->verifyAcl('field', 'update', '');
+		if ($r)
+			return $r;
+
 		$this->view->disable();
 		
 		
@@ -81,6 +93,10 @@ class FieldController extends ControllerBase
 	 */
 	public function queryAction($idCustomField)
 	{
+		$r = $this->verifyAcl('field', 'query', '');
+		if ($r)
+			return $r;
+		
 		$this->view->disable();
 		
 		
@@ -101,6 +117,10 @@ class FieldController extends ControllerBase
 
 	public function newAction($idbase)
 	{
+		$r = $this->verifyAcl('field', 'new', '');
+		if ($r)
+			return $r;
+		
 		$field = new Customfield();
 		$form = new NewFieldForm($field);
 		
@@ -133,6 +153,10 @@ class FieldController extends ControllerBase
 	
 	public function editAction($id)
 	{
+		$r = $this->verifyAcl('field', 'edit', '');
+		if ($r)
+			return $r;
+		
 		$field = new CustomField();
 		$editform = new NewFieldForm($field);
 		$registro = Customfield::findFirstByIdCustomField($id);
@@ -149,6 +173,10 @@ class FieldController extends ControllerBase
 	
 	public function deleteAction($id)
 	{
+		$r = $this->verifyAcl('field', 'delete', '');
+		if ($r)
+			return $r;
+		
 		$registro = Customfield::findFirstByIdCustomField($id);
 		$idbase = $registro->idDbase;
 		$registro->delete();
